pyproject.toml
[project]
name = "temantiket-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Muhammad Iqbal",email = "iqbal.adudu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "langchain[anthropic] (>=0.3.20,<0.4.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-anthropic (>=0.3.9,<0.4.0)",
    "langgraph (>=0.3.5,<0.4.0)"
]

[tool.poetry]
packages = [{include = "temantiket_fastapi", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
